+++ babel.config.js
@@ -1,27 +1,15 @@
 const path = require('path');
 
-const errorCodesPath = path.resolve(__dirname, './docs/public/static/error-codes.json');
-const missingError = process.env.MUI_EXTRACT_ERROR_CODES === 'true' ? 'write' : 'annotate';
-
 function resolveAliasPath(relativeToBabelConf) {
   const resolvedPath = path.relative(process.cwd(), path.resolve(__dirname, relativeToBabelConf));
   return `./${resolvedPath.replace('\\', '/')}`;
 }
-
-const productionPlugins = [
-  ['babel-plugin-react-remove-properties', { properties: ['data-mui-test'] }],
-];
 
 module.exports = function getBabelConfig(api) {
   const useESModules = api.env(['regressions', 'legacy', 'modern', 'stable', 'rollup']);
 
   const defaultAlias = {
     '@pigment-css/react': resolveAliasPath('./packages/pigment-css-react/src'),
-    docs: resolveAliasPath('./docs'),
-    test: resolveAliasPath('./test'),
-    '@mui-internal/api-docs-builder': resolveAliasPath(
-      './node_modules/@mui/monorepo/packages/api-docs-builder',
-    ),
   };
 
   const presets = [
@@ -44,42 +32,8 @@ module.exports = function getBabelConfig(api) {
     '@babel/preset-typescript',
   ];
 
-  const plugins = [
-    [
-      'babel-plugin-macros',
-      {
-        muiError: {
-          errorCodesPath,
-          missingError,
-        },
-      },
-    ],
-    'babel-plugin-optimize-clsx',
-    // Need the following 3 proposals for all targets in .browserslistrc.
-    // With our usage the transpiled loose mode is equivalent to spec mode.
-    ['@babel/plugin-proposal-class-properties', { loose: true }],
-    ['@babel/plugin-proposal-private-methods', { loose: true }],
-    ['@babel/plugin-proposal-private-property-in-object', { loose: true }],
-    ['@babel/plugin-proposal-object-rest-spread', { loose: true }],
-    [
-      '@babel/plugin-transform-runtime',
-      {
-        useESModules,
-        // any package needs to declare 7.4.4 as a runtime dependency. default is ^7.0.0
-        version: '^7.4.4',
-      },
-    ],
-    [
-      'babel-plugin-transform-react-remove-prop-types',
-      {
-        mode: 'unsafe-wrap',
-      },
-    ],
-  ];
+  const plugins = [];
 
-  if (process.env.NODE_ENV === 'production') {
-    plugins.push(...productionPlugins);
-  }
   if (process.env.NODE_ENV === 'test') {
     plugins.push([
       'babel-plugin-module-resolver',
@@ -140,12 +94,6 @@ module.exports = function getBabelConfig(api) {
           ],
         ],
       },
-      legacy: {
-        plugins: [
-          // IE11 support
-          '@babel/plugin-transform-object-assign',
-        ],
-      },
       test: {
         sourceMaps: 'both',
         plugins: [
@@ -160,7 +108,6 @@ module.exports = function getBabelConfig(api) {
       },
       benchmark: {
         plugins: [
-          ...productionPlugins,
           [
             'babel-plugin-module-resolver',
             {