+++ packages/pigment-css-nextjs-plugin/src/virtual-css-loader.js
@@ -2,5 +2,5 @@ export const loader = function virtualFileLoader() {
   const callback = this.async();
   const resourceQuery = this.resourceQuery.slice(1);
   const { source } = JSON.parse(decodeURIComponent(resourceQuery));
-  return callback(null, source);
+  return callback(null, source.replaceAll('__IMP__', '!important'));
 };