+++ packages/pigment-css-react/src/processors/styled.ts
@@ -26,6 +26,7 @@ import type { PluginCustomOptions } from '../utils/cssFnValueToVariable';
 import { cssFnValueToVariable } from '../utils/cssFnValueToVariable';
 import { processCssObject } from '../utils/processCssObject';
 import { valueToLiteral } from '../utils/valueToLiteral';
+import { lowercaseFirstLetter } from '../utils/lowercaseFirstLetter';
 import BaseProcessor from './base-processor';
 import { Primitive, TemplateCallback } from './keyframes';
 import { cache, css } from '../utils/emotion';
@@ -309,8 +310,8 @@ export class StyledProcessor extends BaseProcessor {
    * which we can use to generate our styles.
    * Order of processing styles -
    * 1. CSS directly declared in styled call
-   * 2. CSS declared in theme object's styledOverrides
    * 3. Variants declared in styled call
+   * 2. CSS declared in theme object's styledOverrides
    * 3. Variants declared in theme object
    */
   build(values: ValueCache): void {
@@ -324,11 +325,17 @@ export class StyledProcessor extends BaseProcessor {
     );
     // all the variant definitions are collected here so that we can
     // apply variant styles after base styles for more specific targetting.
-    const variantsAccumulator: VariantData[] = [];
+    let variantsAccumulator: VariantData[] = [];
     (this.styleArgs as ExpressionValue[]).forEach((styleArg) => {
       this.processStyle(values, styleArg, variantsAccumulator, themeImportIdentifier.name);
     });
+    // Generate CSS for default variants first
+    variantsAccumulator.forEach((variant) => {
+      this.processVariant(variant);
+    });
+    variantsAccumulator = [];
     this.processOverrides(values, variantsAccumulator);
+    // Generate CSS for variants declared in `styleOverrides`, if any
     variantsAccumulator.forEach((variant) => {
       this.processVariant(variant);
     });
@@ -470,9 +477,8 @@ export class StyledProcessor extends BaseProcessor {
       if (!overrides) {
         return;
       }
-      const overrideStyle = (overrides[value.slot.toLowerCase()] || overrides[value.slot]) as
-        | string
-        | CSSObject;
+      const overrideStyle = (overrides[lowercaseFirstLetter(value.slot)] ||
+        overrides[value.slot]) as string | CSSObject;
       const className = this.getClassName();
       if (typeof overrideStyle === 'string') {
         this.collectedStyles.push([className, overrideStyle, null]);
@@ -490,7 +496,7 @@ export class StyledProcessor extends BaseProcessor {
     if (
       'variants' in componentData &&
       componentData.variants &&
-      value.slot.toLowerCase() === 'root'
+      lowercaseFirstLetter(value.slot) === 'root'
     ) {
       variantsAccumulator.push(...(componentData.variants as unknown as VariantData[]));
     }