+++ packages/pigment-css-react/src/styled.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-underscore-dangle */
 import * as React from 'react';
 import clsx from 'clsx';
 import isPropValid from '@emotion/is-prop-valid';
@@ -26,7 +27,7 @@ function isHtmlTag(tag) {
   );
 }
 
-const slotShouldForwardProp = (key) => key !== 'sx' && key !== 'as' && key !== 'ownerState';
+const slotShouldForwardProp = (key) => key !== 'as' && key !== 'ownerState';
 const rootShouldForwardProp = (key) => slotShouldForwardProp(key) && key !== 'classes';
 
 /**
@@ -58,7 +59,12 @@ export default function styled(tag, componentMeta = {}) {
       finalShouldForwardProp = slotShouldForwardProp;
     }
   }
-  const shouldUseAs = !finalShouldForwardProp('as');
+  let shouldUseAs = !finalShouldForwardProp('as');
+  if (typeof tag !== 'string' && tag.__styled_by_pigment_css) {
+    // If the tag is a Pigment styled component,
+    // render the styled component and pass the `as` prop down
+    shouldUseAs = false;
+  }
   /**
    * This is the runtime `styled` function that finally renders the component
    * after transpilation through WyW-in-JS. It makes sure to add the base classes,
@@ -80,11 +86,11 @@ export default function styled(tag, componentMeta = {}) {
     const { displayName, classes = [], vars: cssVars = {}, variants = [] } = options;
 
     const StyledComponent = React.forwardRef(function StyledComponent(inProps, ref) {
-      const { as, className, sx, style, ownerState, ...props } = inProps;
-      const Component = (shouldUseAs && as) || tag;
+      const { className, sx, style, ownerState, ...props } = inProps;
+      const Component = (shouldUseAs && inProps.as) || tag;
       const varStyles = Object.entries(cssVars).reduce(
         (acc, [cssVariable, [variableFunction, isUnitLess]]) => {
-          const value = variableFunction(props);
+          const value = variableFunction(inProps);
           if (typeof value === 'undefined') {
             return acc;
           }
@@ -97,25 +103,19 @@ export default function styled(tag, componentMeta = {}) {
         },
         {},
       );
-      const sxClass = typeof sx === 'string' ? sx : sx?.className;
-      const sxVars = sx && typeof sx !== 'string' ? sx.vars : undefined;
 
-      if (sxVars) {
-        Object.entries(sxVars).forEach(([cssVariable, [value, isUnitLess]]) => {
-          if (typeof value === 'string' || isUnitLess) {
-            varStyles[`--${cssVariable}`] = value;
+      const finalClassName = clsx(classes, className, getVariantClasses(inProps, variants));
+
+      if (inProps.as && !shouldForwardProp) {
+        // Reassign `shouldForwardProp` if incoming `as` prop is a React component
+        if (!isHtmlTag(Component)) {
+          if (slot === 'Root' || slot === 'root') {
+            finalShouldForwardProp = rootShouldForwardProp;
           } else {
-            varStyles[`--${cssVariable}`] = `${value}px`;
+            finalShouldForwardProp = slotShouldForwardProp;
           }
-        });
+        }
       }
-
-      const finalClassName = clsx(
-        classes,
-        sxClass,
-        className,
-        getVariantClasses(inProps, variants),
-      );
 
       const newProps = {};
       // eslint-disable-next-line no-restricted-syntax
@@ -124,7 +124,7 @@ export default function styled(tag, componentMeta = {}) {
           continue;
         }
 
-        if (finalShouldForwardProp(key)) {
+        if (finalShouldForwardProp(key) || (!shouldUseAs && key === 'as')) {
           newProps[key] = props[key];
         }
       }
@@ -133,7 +133,6 @@ export default function styled(tag, componentMeta = {}) {
         <Component
           {...newProps}
           // pass down `ownerState` to nested styled components
-          // eslint-disable-next-line no-underscore-dangle
           {...(Component.__styled_by_pigment_css && { ownerState })}
           ref={ref}
           className={finalClassName}
@@ -150,7 +149,6 @@ export default function styled(tag, componentMeta = {}) {
       componentName = `${name}${slot ? `-${slot}` : ''}`;
     }
     StyledComponent.displayName = `Styled(${componentName})`;
-    // eslint-disable-next-line no-underscore-dangle
     StyledComponent.__styled_by_pigment_css = true;
 
     return StyledComponent;