+++ packages/pigment-css-react/src/sx.d.ts
@@ -3,4 +3,4 @@ import type { ThemeArgs } from './theme';
 
 export type SxProp = CSSObjectNoCallback | ((themeArgs: ThemeArgs) => CSSObjectNoCallback);
 
-export default function sx(arg: SxProp, componentClass?: string): string;
+export default function sx(arg: SxProp | Array<SxProp>, componentClass?: string): string;