+++ packages/pigment-css-react/tests/sx/sx.test.ts
@@ -46,12 +46,68 @@ describe('Pigment CSS - sx prop', () => {
     expect(output.css).to.equal(fixture.css);
   });
 
+  it('jsx sx-prop with logical and conditional expressions', async () => {
+    const { output, fixture } = await runTransformation(
+      path.join(__dirname, 'fixtures/sx-jsx.input.js'),
+      {
+        themeArgs: {
+          theme,
+        },
+      },
+    );
+
+    expect(output.js).to.equal(fixture.js);
+    expect(output.css).to.equal(fixture.css);
+  });
+
   it('sx prop with theme spread', async () => {
     const { output, fixture } = await runTransformation(
       path.join(__dirname, 'fixtures/sxProps2.input.js'),
       {
         themeArgs: {
           theme,
+        },
+      },
+    );
+
+    expect(output.js).to.equal(fixture.js);
+    expect(output.css).to.equal(fixture.css);
+  });
+
+  it('sx prop support array', async () => {
+    const { output, fixture } = await runTransformation(
+      path.join(__dirname, 'fixtures/sx-array.input.js'),
+      {
+        themeArgs: {
+          theme,
+        },
+      },
+    );
+
+    expect(output.js).to.equal(fixture.js);
+    expect(output.css).to.equal(fixture.css);
+  });
+
+  it('sx prop shorthand', async () => {
+    const { output, fixture } = await runTransformation(
+      path.join(__dirname, 'fixtures/sx-shorthand.input.js'),
+      {
+        themeArgs: {
+          theme: {
+            cssVarPrefix: 'mui',
+            palette: {
+              grey: {
+                50: '#FBFCFE',
+              },
+            },
+            vars: {
+              palette: {
+                grey: {
+                  50: 'var(--mui-palette-grey-50)',
+                },
+              },
+            },
+          },
         },
       },
     );