+++ CHANGELOG.md
@@ -1,5 +1,46 @@
 # [Versions](https://mui.com/versions/)
 
+## v0.0.11
+
+_May 8, 2024_
+
+A big thanks to the 2 contributors who made this release possible. Here are some highlights ✨:
+
+- ✨ `sx` prop is now working on any components (#37) @siriwatknp
+
+### `@pigment-css/react@0.0.11`
+
+- Support array for `sx` prop (#52) @siriwatknp
+- Replace `sx` callback object param with `theme` (#51) @siriwatknp
+- Enable `sx` prop on any components (#37) @siriwatknp
+- Replace lowercase with lowerFirst (#54) @siriwatknp
+- Reorder the css generation order for styled calls (#56) @brijeshb42
+
+### Core
+
+- Set root version to 0.0.10 (#44) @siriwatknp
+
+All contributors of this release in alphabetical order: @@brijeshb42, @siriwatknp
+
+## v0.0.10
+
+<!-- generated comparing v0.0.9..master -->
+
+_May 1, 2024_
+
+A big thanks to the 2 contributors who made this release possible.
+
+### `@pigment-css/react@0.0.10`
+
+- Fix styled inheritance (#40) @siriwatknp
+- Fix prop forwarding when `as` component is provided to HTML styled component (#43) @siriwatknp
+
+### `@pigment-css/nextjs-plugin@0.0.10`
+
+- Transform `!important` CSS to an intermediate representation (#38) @brijeshb42
+
+All contributors of this release in alphabetical order: @brijeshb42, @siriwatknp
+
 ## v0.0.9
 
 <!-- generated comparing v0.0.8..master -->